+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log (go-package-manager)
 
+## 0.11.0
+
+- feat: load `.env` files from `$HOME/.gpm` and project directories, if exist, automatically
+- feat: add `--env-file` file to load environment variables from external variables
+- fix: exit app if special files could not be loaded
+
 ## 0.10.1
 
 - initial release